+++ development_plan.md
@@ -53,3 +53,6 @@
 
 ### Main Functions:
 - Modify data_ingestor script to align with new changes
+
+### Tests:
+- write tests for individual functions in data_processing module
